Significantly reduce memory use by (correctly) using uint32 instead of uint64 for memory addresses.  Also save off start and end addresses, rather than computing them from subsection indices every time.  Fix a bug introduced when memory_manager changed.

+++ momo_modules/mainboard/src/momo/sensor_event_log.h
@@ -12,8 +12,8 @@ typedef struct {
 
 void init_sensor_event_log( uint8 start_subsector, uint8 num_subsectors );
 bool log_sensor_event( uint8 stream_id, uint8 meta, const rtcc_datetime* timestamp, uint64 *value );
-unsigned int read_sensor_events( sensor_event* events, uint64 max );
+uint32 read_sensor_events( sensor_event* events, uint32 max );
 bool sensor_event_log_empty();
-unsigned long sensor_event_log_count();
+uint32 sensor_event_log_count();
 
 #endif
